squash me:
* move to long instead of double
* integrate into Spring Boot Actuator if enabled

+++ engine/src/main/java/org/camunda/bpm/engine/impl/metrics/CounterMetricSampleFamily.java
@@ -30,7 +30,7 @@ public class CounterMetricSampleFamily extends MetricSampleFamily {
   }
 
   @Override
-  public CounterMetricSampleFamily addValue(List<String> labelsIn, double value) {
+  public CounterMetricSampleFamily addValue(List<String> labelsIn, long value) {
     List<String> labels = labelsIn == null ? Collections.emptyList() : labelsIn;
     if (labels.size() != labelNames.size()) {
       throw new IllegalArgumentException("Incorrect number of labels.");
